feat: Add business logic to not allow user to delete ongoing task

+++ src/controllers/admin/tasks.js
@@ -1,5 +1,6 @@
 const joi = require("joi");
 const {
+  appError,
   notFoundError,
   validationError,
   apiResponse,
@@ -131,6 +132,8 @@ exports.delete = async (req, res) => {
     const id = parseInt(req.params.id);
     const task = await TaskDAO.fetchOne({ id });
     if (!task) throw notFoundError();
+    if (task.status === "IN_PROGRESS")
+      throw appError("Task cannot be deleted when it is in progress");
     await TaskDAO.delete({ id });
     return apiResponse(res, {
       success: true,
